refactor: dedup test fixtures, split process/styles/sound modules
Cleanup-pass findings #10, #16, #12, #13 from the code-quality scan.
No behavior changes; all tests still pass.

- tests/common/mod.rs: shared setup_temp_home/set_temp_home, replacing
  eight identical definitions across integration test files.
- src/process/: kill_with_fallback + process_exists moved to the shared
  mod.rs (using portable nix kill(None) for existence). Each platform
  module now only exposes collect_pid_tree (Linux via /proc, macOS via
  ps) plus get_foreground_pid.
- src/tui/styles/: split the 857-line file into palette.rs (RGB->xterm
  palette mapping), themes.rs (Theme struct + built-in constructors +
  hex_color serde helper), and mod.rs (loading/discovery glue).
- src/sound/: split the 717-line file into config.rs (SoundConfig +
  volume helpers), discovery.rs (sounds dir probing), bundled.rs
  (GitHub installer), playback.rs (afplay/paplay/aplay dispatch), and
  mod.rs (transition-to-sound glue).
