+++ src/process/linux.rs
@@ -3,41 +3,13 @@
 use std::fs;
 use std::path::Path;
 
-use nix::sys::signal::{kill, Signal};
-use nix::unistd::Pid;
-use tracing::debug;
-
-/// Kill a process and all its descendants
-/// Uses SIGTERM first, then SIGKILL after a short delay for stragglers
-pub fn kill_process_tree(pid: u32) {
-    // Collect all descendant PIDs first (children, grandchildren, etc.)
-    let mut pids_to_kill = vec![pid];
-    collect_descendants(pid, &mut pids_to_kill);
-
-    debug!(
-        pid,
-        descendants = ?pids_to_kill,
-        "Killing process tree"
-    );
-
-    // Kill in reverse order (children first, then parent) with SIGTERM
-    for &p in pids_to_kill.iter().rev() {
-        let _ = kill(Pid::from_raw(p as i32), Signal::SIGTERM);
-    }
-
-    // Brief pause to let processes handle SIGTERM gracefully
-    std::thread::sleep(std::time::Duration::from_millis(100));
-
-    // SIGKILL any survivors
-    for &p in pids_to_kill.iter().rev() {
-        if process_exists(p) {
-            debug!(pid = p, "Process survived SIGTERM, sending SIGKILL");
-            let _ = kill(Pid::from_raw(p as i32), Signal::SIGKILL);
-        }
-    }
+/// Collect `pid` and every descendant by walking `/proc/<pid>/stat`.
+pub(super) fn collect_pid_tree(pid: u32) -> Vec<u32> {
+    let mut pids = vec![pid];
+    collect_descendants(pid, &mut pids);
+    pids
 }
 
-/// Recursively collect all descendant PIDs of a process
 fn collect_descendants(pid: u32, pids: &mut Vec<u32>) {
     let proc_dir = Path::new("/proc");
     if !proc_dir.exists() {
@@ -71,11 +43,6 @@ fn collect_descendants(pid: u32, pids: &mut Vec<u32>) {
             }
         }
     }
-}
-
-/// Check if a process still exists
-fn process_exists(pid: u32) -> bool {
-    Path::new(&format!("/proc/{}", pid)).exists()
 }
 
 /// Get the foreground process group leader for a shell PID
